fix: resolving issue with scopes
